Move all files for correct location for webpacker
