+++ app/javascript/src/file_tree/file_tree_events.js
@@ -2,7 +2,7 @@
  * Created by lobashov-2 on 18.05.17.
  */
 
-function setEventToOpenFolder() {
+window.setEventToOpenFolder = function() {
     $(".folder-name").on('click', function () {
         var currentDisplay = $(this).next(".folder-inside").css("display");
         var icon = $(this).children("i");
@@ -19,7 +19,7 @@ function setEventToOpenFolder() {
     });
 }
 
-function eventToAddTestInQueue(elem) {
+window.eventToAddTestInQueue = function(elem) {
     elem.on('click', function(){
         addTestInQueue($(this).attr('full-path'), getDocBranch(), getSelectedPortalUrl());
         getUpdatedDataFromServer();
@@ -27,7 +27,7 @@ function eventToAddTestInQueue(elem) {
     });
 }
 
-function addFolderInQueue(folder_elem) {
+window.addFolderInQueue = function(folder_elem) {
     var tests = [];
     folder_elem.find('.add-button-file').each(function(){
         tests.push($(this).attr('full-path'));
@@ -40,7 +40,7 @@ function addFolderInQueue(folder_elem) {
 
 }
 
-function eventToAddFolderInQueue(folder_elem) {
+window.eventToAddFolderInQueue = function(folder_elem) {
     folder_elem.on('click', function(){
         addFolderInQueue($(this).parent());
         getUpdatedDataFromServer();
