+++ app/javascript/src/server_info.js
@@ -1,32 +1,32 @@
 /**
  * Created by Pavel.Lobashov on 14.10.15.
  */
-fetch_server_ip = function (server) {
+window.fetch_server_ip = function(server) {
     $.ajax({
         url: '/servers/cloud_server_fetch_ip',
         type: 'GET',
         data: {
             'server': server
         },
-        beforeSend: function () {
+        beforeSend: function() {
             showOverlay('Fetching current server IP');
         },
-        complete: function () {
+        complete: function() {
             hideOverlay();
         },
-        success: function (data) {
+        success: function(data) {
             update_ip_value(data.ip);
         },
-        error: function (e) {
+        error: function(e) {
             console.log(e.message);
         }
     });
 };
 
-update_ip_value = function(value) {
+window.update_ip_value = function(value) {
         $("#server_address").val(value);
 };
 
-function fetchCurrentServerIp() {
+window.fetchCurrentServerIp = function() {
     fetch_server_ip($( "#server_name" ).val());
 }
