+++ app/javascript/src/server_log.js
@@ -2,7 +2,7 @@
  * Created by lobashov-2 on 12.05.17.
  */
 
-function fetch_server_log(server_name) {
+window.fetch_server_log = function(server_name) {
     $.ajax({
         url: 'servers/log',
         data: { server: server_name },
@@ -17,17 +17,17 @@ function fetch_server_log(server_name) {
     });
 }
 
-function fill_server_log(server_name, log) {
+window.fill_server_log = function(server_name, log) {
     var log_div = $("#" + server_name + ' .log');
     log_div.text(log);
 }
 
-function empty_server_log(server_name) {
+window.empty_server_log = function(server_name) {
     var log_div = $("#" + server_name + ' .log');
     log_div.text('');
 }
 
-function server_log_visible(server_name) {
+window.server_log_visible = function(server_name) {
     var log_div = $("#" + server_name + ' .log');
     return log_div.is(':visible');
 }
