+++ app/javascript/src/spec_browsers.js
@@ -1,6 +1,6 @@
 // Place all the behaviors and hooks related to the matching controller here.
 // All this logic will automatically be available in application.js.
 
-function getSpecBrowser() {
+window.getSpecBrowser = function() {
     return $('#browser_0').val();
 }
