+++ app/javascript/src/spec_languages.js
@@ -1,6 +1,6 @@
 // Place all the behaviors and hooks related to the matching controller here.
 // All this logic will automatically be available in application.js.
 
-function getSpecLanguage() {
+window.getSpecLanguage = function() {
     return $('#language_0').val();
 }
