+++ app/javascript/src/test_list.js
@@ -2,7 +2,7 @@
  * Created by lobashov-2 on 17.05.17.
  */
 
-function fullPathForFile(file) {
+window.fullPathForFile = function(file) {
     var full_path = file.attr('data-test');
     var parent_folder = file.get(0).parentNode.parentNode.parentNode.children[0];
     while (parent_folder.hasAttribute('data-test')) {
@@ -12,14 +12,14 @@ function fullPathForFile(file) {
     return full_path;
 }
 
-function addFullPaths(root_folder) {
+window.addFullPaths = function(root_folder) {
     var project_folder = projectFolder();
     root_folder.find('.add-button-file').each(function () {
         $(this).attr('full-path', '~/RubymineProjects/' + project_folder + '/' + fullPathForFile($(this)));
     });
 }
 
-function projectFolder() {
+window.projectFolder = function() {
     var project = activeProject();
     if (project == 'ONLYOFFICE/testing-documentserver') {
         return 'OnlineDocuments';
