+++ app/views/runner/index.html.erb
@@ -1,5 +1,6 @@
 <% content_for :head do %>
   <script type="text/javascript">
+      let myRunner = new Runner();
       $(document).ready(function () {
           myRunner.showServers();
           myRunner.eventToSortTestQueue();
